requirements.txt
nltk
streamlit>=1.42.0
pysqlite3-binary
chromadb>=0.6.3
llama-index>=0.12.19
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
google-generativeai
sentence-transformers
torch>=2.1.0
transformers>=4.36.0
protobuf>=4.25.1
nest-asyncio>=1.5.6
firebase-admin==6.4.0
pyrebase4
pytz
schedule
spacy
yake